requirements.txt
pygame
python-dotenv
langchain
langchain-google-genai
pyperclip